app/requirements.txt
blinker==1.8.2
branca==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.9
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
folium==0.16.0
fonttools==4.52.1
fpdf==1.7.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.3.6
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.6
Werkzeug==3.0.4
xyzservices==2024.4.0
Flask-SocketIO==5.5.1
paho-mqtt==2.1.0
gunicorn==23.0.0